Refactor Galaxy testing stuff toward greater reuse.
