+++ planemo/galaxy_test/__init__.py
@@ -1,0 +1,3 @@
+from .actions import run_in_config
+
+__all__ = ["run_in_config"]
