+++ tests/test_galaxy_test.py
@@ -4,7 +4,7 @@ import os
 
 from .test_utils import TEST_DATA_DIR
 
-from planemo import galaxy_test
+from planemo.galaxy_test import structures
 
 nose_1_3_report = os.path.join(TEST_DATA_DIR, "xunit_nose_1_3.xml")
 nose_0_11_report = os.path.join(TEST_DATA_DIR, "xunit_nose_0_11.xml")
@@ -19,10 +19,10 @@ def get_test_id_old():
 
 
 def _get_test_id(path):
-    xml_tree = galaxy_test.parse_xunit_report(path)
+    xml_tree = structures.parse_xunit_report(path)
     root = xml_tree.getroot()
-    first_testcase = galaxy_test.find_cases(root)[0]
-    test_id = galaxy_test.case_id(first_testcase)
+    first_testcase = structures.find_cases(root)[0]
+    test_id = structures.case_id(first_testcase)
     assert test_id.label == "cat[0]"
     expected_id = "functional.test_toolbox.TestForTool_cat.test_tool_000000"
     assert test_id.id == expected_id
